oops, bug in little-endian optimization
git-svn-id: https://protobuf-c.googlecode.com/svn/trunk@70 00440858-1255-0410-a3e6-75ea37f81c3a

+++ src/google/protobuf-c/protobuf-c.c
@@ -1070,7 +1070,7 @@ static inline uint64_t
 parse_fixed_uint64 (const uint8_t *data)
 {
 #if IS_LITTLE_ENDIAN
-  uint32_t t;
+  uint64_t t;
   memcpy (&t, data, 8);
   return t;
 #else
